+++ CHANGELOG.md
@@ -1,5 +1,41 @@
 # Changelog
 
+## [1.6.53] - 2025-11-12
+
+### Beta
+- Version 1.6.53 release
+
+### Technical Details
+- Version: 1.6.53
+- Release Type: beta
+- Build Number: 202511122104
+- Release Date: 2025-11-12T21:04:40.518Z
+
+
+## [1.6.52] - 2025-11-12
+
+### Beta
+- Version 1.6.52 release
+
+### Technical Details
+- Version: 1.6.52
+- Release Type: beta
+- Build Number: 202511122042
+- Release Date: 2025-11-12T20:42:13.512Z
+
+
+## [1.6.51] - 2025-11-12
+
+### Beta
+- Version 1.6.51 release
+
+### Technical Details
+- Version: 1.6.51
+- Release Type: beta
+- Build Number: 202511122033
+- Release Date: 2025-11-12T20:33:43.475Z
+
+
 ## [1.6.50] - 2025-11-12
 
 ### Bugfixes